+++ lib/src/core/l10n/arb/app_localizations_ar.dart
@@ -5,8 +5,8 @@ import 'app_localizations.dart';
 // ignore_for_file: type=lint
 
 /// The translations for Arabic (`ar`).
-class AppLocalizationsAr extends AppLocalizations {
-  AppLocalizationsAr([String locale = 'ar']) : super(locale);
+class FrostedAppLocalizationsAr extends FrostedAppLocalizations {
+  FrostedAppLocalizationsAr([String locale = 'ar']) : super(locale);
 
   @override
   String get signIn => 'تسجيل الدخول';