+++ lib/src/core/l10n/arb/app_localizations_en.dart
@@ -5,8 +5,8 @@ import 'app_localizations.dart';
 // ignore_for_file: type=lint
 
 /// The translations for English (`en`).
-class AppLocalizationsEn extends AppLocalizations {
-  AppLocalizationsEn([String locale = 'en']) : super(locale);
+class FrostedAppLocalizationsEn extends FrostedAppLocalizations {
+  FrostedAppLocalizationsEn([String locale = 'en']) : super(locale);
 
   @override
   String get signIn => 'Sign In';