+++ lib/src/features/auth/presentation/widgets/forgot_password_form_widget.dart
@@ -57,7 +57,7 @@ class _ForgotPasswordFormWidgetState extends State<ForgotPasswordFormWidget> {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
-    final l10n = AppLocalizations.of(context)!;
+    final l10n = FrostedAppLocalizations.of(context)!;
 
     return Form(
       key: _formKey,