+++ lib/src/features/auth/presentation/widgets/verify_email_form_widget.dart
@@ -48,7 +48,7 @@ class _VerifyEmailFormWidgetState extends State<VerifyEmailFormWidget> {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
-    final l10n = AppLocalizations.of(context)!;
+    final l10n = FrostedAppLocalizations.of(context)!;
 
     return Form(
       key: _formKey,